pyproject.toml
[project]
name = "amazonscrapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.12",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.11",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tinydb>=4.8.2",
]
